safari-600.7-branch/LayoutTests/inspector-obsolete/timeline/timeline-load-expected.txt

Tests the Timeline save/load functionality.

Also tests save/load of unknown (incompatible) record types. (Bug 91714)


Running: testNormal
passed

Running: testIncompatible
passed

Running: testMalformed
passed
